LPS-110819 Moves siteNavigationMenuNames to react context
